Allow runtime configuration of MasterProxy
By not starting up an application, and instead having the user define a
proxy module to add to their supervision tree (similar to how a Phoenix
Endpoint is added to the user's supervision tree) it becomes easier for
the user to define the backends for MasterProxy at runtime.

In particular I need this for using MasterProxy with SiteEncrypt since
SiteEncrypt generates its configuration at runtime.